Linux warning fixes
Review URL: https://codereview.appspot.com/6500045/



git-svn-id: http://skia.googlecode.com/svn/trunk@5337 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/image/SkImage_Raster.cpp
@@ -43,7 +43,7 @@ public:
         }
 
         int64_t size = (int64_t)info.fHeight * rowBytes;
-        if (size > kMaxPixelByteSize) {
+        if (size > (int64_t)kMaxPixelByteSize) {
             return false;
         }
         return true;
